% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_get}
\alias{cdb_api_get}
\title{Make a HTTP GET call to the CartoDB map API}
\usage{
cdb_api_get(endpoint, version = "v1", query = list(), ...)
}
\description{
Make a HTTP GET call to the CartoDB map API
}

