% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_post}
\alias{cdb_api_post}
\title{Make a POST call to the CartoDB map API}
\usage{
cdb_api_post(endpoint, version = "v1", body = NULL, query = NULL,
  encode = "json", ...)
}
\description{
Make a POST call to the CartoDB map API
}

