% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_put}
\alias{cdb_api_put}
\title{Make a PUT call to the CartoDB map API}
\usage{
cdb_api_put(endpoint, version = "v1", encode = "json", ...)
}
\description{
Make a PUT call to the CartoDB map API
}

