% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_import}
\alias{cdb_import}
\title{Import file into CartoDB}
\usage{
cdb_import(url, api_key, server = "cartodb.com")
}
\value{
Response from POST request
}
\description{
Import a file into a CartoDB database.
}

