% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_import_file}
\alias{cdb_import_file}
\title{Import data from local file into CartoDB}
\usage{
cdb_import_file(path, user, api_key, privacy = "private", ...)
}
\value{
Response from POST request
}
\description{
Import data from local file into CartoDB
}

