% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_import_url}
\alias{cdb_import_url}
\title{Import data from URL into CartoDB}
\usage{
cdb_import_url(url, user, api_key, privacy = "private", ...)
}
\value{
Response from POST request
}
\description{
Import data from URL into CartoDB
}

