% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_synchronizations}
\alias{cdb_synchronizations}
\title{List sync tables for a given CartoDB account}
\usage{
cdb_synchronizations(...)
}
\value{
Response from GET request
}
\description{
List sync tables for a given CartoDB account
}

