% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_synctable_check}
\alias{cdb_synctable_check}
\title{Check if a table is a sync table}
\usage{
cdb_synctable_check(table, user, api_key, ...)
}
\value{
Response from POST request
}
\description{
Check if a table is a sync table
}

