% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_synctable_convert}
\alias{cdb_synctable_convert}
\title{Convert sync table to regular table (remove sync)}
\usage{
cdb_synctable_convert(table)
}
\value{
Response from DELETE request
}
\description{
Convert sync table to regular table (remove sync)
}

