% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_synctable_getid}
\alias{cdb_synctable_getid}
\title{Get id for a sync table}
\usage{
cdb_synctable_getid(table, user, api_key, ...)
}
\value{
Response from POST request
}
\description{
Get id for a sync table
}

