% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_synctable_import}
\alias{cdb_synctable_import}
\title{Import file into CartoDB as a synced table}
\usage{
cdb_synctable_import(data_url, user, api_key, sync_interval = 604800, ...)
}
\value{
response from POST request
}
\description{
Import file into CartoDB as a synced table
}

