% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_synctable_status}
\alias{cdb_synctable_status}
\title{Check status of sync table}
\usage{
cdb_synctable_status(table, user, api_key, ...)
}
\value{
sync table status
}
\description{
Check status of sync table
}

