% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_import.R
\name{cdb_table_exists}
\alias{cdb_table_exists}
\title{Check if a CartoDB table exists}
\usage{
cdb_table_exists(table)
}
\value{
Response from POST request
}
\description{
Check if a CartoDB table exists
}

