% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_map.R
\name{create_anon_map}
\alias{create_anon_map}
\title{Instantiate an anonymous map}
\usage{
create_anon_map(layers, user, api_key, raster = FALSE, ...)
}
\value{
response from POST request
}
\description{
Instantiate an anonymous map
}

