% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_map.R
\name{create_named_map}
\alias{create_named_map}
\title{Create a named map}
\usage{
create_named_map(name, layers, user, api_key, overwrite = FALSE, ...)
}
\value{
response from POST request
}
\description{
Create a named map
}

