% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_map.R
\name{list_named_maps}
\alias{list_named_maps}
\title{List named maps for a given CartoDB account}
\usage{
list_named_maps(...)
}
\value{
Response from GET request
}
\description{
List named maps for a given CartoDB account
}

