% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_map.R
\name{mapconfig_raster}
\alias{mapconfig_raster}
\title{Setup mapconfig for a raster layer}
\usage{
mapconfig_raster(sql, cartocss, interactivity = NA, raster_band = 1,
  type = "cartodb", cartocss_version = "2.3.0")
}
\value{
response from POST request
}
\description{
Setup mapconfig for a raster layer
}

