% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdb_map.R
\name{mapconfig_vector}
\alias{mapconfig_vector}
\title{Setup mapconfig for a vector layer}
\usage{
mapconfig_vector(sql, cartocss, interactivity = NA, raster_band = 1,
  type = "cartodb", cartocss_version = "2.3.0")
}
\value{
vector layer mapconfig
}
\description{
Setup mapconfig for a vector layer
}

