% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_delete}
\alias{cdb_api_delete}
\title{Send a DELETE request to a CartoDB API}
\usage{
cdb_api_delete(endpoint, version = "v1", encode = "json", ...)
}
\description{
Send a DELETE request to a CartoDB API
}

