% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_delete}
\alias{cdb_api_delete}
\title{Make a POST call to the CartoDB map API}
\usage{
cdb_api_delete(user, endpoint, domain = "cartodb.com", api_key = "",
  version = "v1", protocol = "https", subdomain = TRUE, encode = "json",
  ...)
}
\description{
Make a POST call to the CartoDB map API
}

