% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_post}
\alias{cdb_api_post}
\title{Send a POST request to a CartoDB API}
\usage{
cdb_api_post(endpoint, version = "v1", body = NULL, query = NULL,
  encode = "json", ...)
}
\description{
Send a POST request to a CartoDB API
}

