% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{cdb_api_post}
\alias{cdb_api_post}
\title{Make a POST call to the CartoDB map API}
\usage{
cdb_api_post(user, endpoint, body, domain = "cartodb.com", query = list(),
  version = "v1", protocol = "https", subdomain = TRUE, encode = "json",
  ...)
}
\description{
Make a POST call to the CartoDB map API
}

